recipes-core/images: Modify the style of luv_netconsole
Modify both luv-live-image.bb and luv-netboot-image.bb files to reflect
the change in style of the kernel parameter luv_netconsole.

The change in style is to match with the new configuration file in luv
'luv.cfg' and since we are decoupling from the grub.cfg, it makes
sense to prepend the variables with 'LUVCFG_' instead of 'APPEND_'.

Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>