initscripts/: Add a new file 'save_luv_netboot_results' and then add it to bin directory
save_luv_netboot_results is a script that sends the
luv-netboot-image results in a tar ball to an URL, that is specified in
the kernel command line.

Also modify the initscripts_1.0.bbappend to include curl utility
and to install save_luv_netboot_results in bin directory. By having it
in the bin directory, this script need not run at any level(when called
by init)and can be called/invoked by luv-test-manager.

Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>