meta-luv/utils/: Add a new file "modify_luv_netboot_efi.py" to luv
As there is no way to modify the EFI application to set the variables
like luv-netconsole and luv_storage to use network debugging feature
and to submit the results to a webserver respectively, we need a script
that lets user modify the parameters.

By modifying such parameters, it is easy to configure the features in
LUV while booting via netboot(luv-netboot-image.efi)

The script file will be available for download and it is not included
in the image itself.

Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>