systemd : Enable networking
After shifting to systemD, the LUV image is not being assigned an
IP on boot. This is because it is still using a systemV startup script
to do the same. Since systemD names its interfaces differently, we could
not see any interfaces with a valid IP. This patch introduces a service
file, networking.service which will bring up the network and make sure
an IP is assigned during boot.

Since systemd has a different naming convention for the interfaces, we
have updated the interface name from eth0/1 to eno1/enp1s0f0/enp1s0f1.

Right now, we wait for the eno1 device to come up before the networking
service file can run, but we have to come up with a more generic solution
because not all systems will have a interface named eno1.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>