luv : Enable telemetrics in LUV
LUV(Linux UEFI Validation) brings together multiple separate upstream
test suites into a cohesive and easy-to-use product and validates UEFI
firmware at critical levels of the software stack. It may be possible
that one of the test suites crashes while running. It may be even
possible that a kernel panic is observed. Under these scenarios, it
would be useful for LUV to call home and submit forensic data to analyze
and address the problem. Of course, this will be an opt-in feature and
users will get clear indication that data is being collected.
We have used the telemetrics-client code developed by the clear-linux
team and tried to incorporate it in LUV. It has support for crashprobe,
oopsprobe and pstore-probe.

This patch consists of:

1. A yocto recipe which fetches the clear-linux telemetrics-client repo,
builds it and installs all the necessary service files, daemons and
probes.

2. The telemetrics-client requires the VERSION_ID to be a number, else
records are not transmitted to the telemetrics server. Currently,
VERSION_ID provided by LUV has some special symbols; hence the records
do not reach the server. We have introduced a new BUILD_ID in the
os-release recipe and given it a random numeric value which will replace
the VERSION_ID. Later, the telemetrics server can use this ID to filter
out the records generated by LUV.

3. Add the necessary kernel config options to enable the pstore file
system.

4. Increase the crash kernel memory and allocate it in high memory.
Currently, the Crash kernel allocated is 256 MB which is very less to
store the reboot kernel. Also, a contiguous larger memory(512MB) is not
found always in the normal memory, hence we allocate it in the high
memory region.

5. Add the crash_kexec_post_notifiers option to make sure that kdump
runs first after the crash followed by kexec. Hence, on a crash, the
dump is first stored and then kexec takes over to reboot the system.

6. Add a kernel line parameter which lets the user opt-in to the
telemetrics feature. By default, this feature is diabled.

Currently, this feature is only supported for x86 systems.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>